move TaskWithProgress to separate module
