+++ ewokscore/__init__.py
@@ -1,4 +1,5 @@
 from .task import Task  # noqa: F401
+from .taskwithprogress import TaskWithProgress  # noqa: F401
 from .graph import load_graph  # noqa: F401
 from .graph import execute_graph  # noqa: F401
