+++ ewokscore/task.py
@@ -4,8 +4,6 @@ from .variable import VariableContainer
 from .variable import VariableContainerNamespace
 from .variable import ReadOnlyVariableContainerNamespace
 from .registration import Registered
-from .progress import BaseProgress
-from typing import Union
 
 
 class TaskInputError(ValueError):
@@ -254,29 +252,3 @@ class Task(Registered, hashing.UniversalHashable, register=False):
     def run(self):
         """To be implemented by the derived classes"""
         raise NotImplementedError
-
-
-class TaskWithProgress(Task):
-    """
-    Task within a progress to display task advancement
-    """
-
-    def __init__(
-        self, inputs=None, varinfo=None, progress: Union[None, BaseProgress] = None
-    ):
-        super().__init__(inputs=inputs, varinfo=varinfo)
-        self._task_progress = progress
-
-    @property
-    def progress(self) -> Union[None, int]:
-        """Task advancement. If a task progress is not provided then return
-        None"""
-        if self._task_progress is not None:
-            return self._task_progress.progress
-        else:
-            return None
-
-    @progress.setter
-    def progress(self, progress: int):
-        if self._task_progress:
-            self._task_progress.progress = progress
